Create separate files for vertical menu stories
