+++ stories/Menu/index.stories.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Icon, Menu } from 'semantic-ui-react'
+import { Menu } from 'semantic-ui-react'
 import { storiesOf } from '@storybook/react'
 
 const items = [
@@ -15,55 +15,6 @@ const items = [
 ];
 
 storiesOf('Menu', module)
-  .add('vertical', () => {
-    const items = [
-      {
-        content: 'Header',
-        header: true
-      },
-      {
-        content: 'Home',
-        icon: 'home'
-      },
-      {
-        content: 'Places',
-        icon: 'place'
-      },
-      {
-        content: 'Reports',
-        icon: 'assignment'
-      }
-    ]
-    return (
-      <React.Fragment>
-        <Menu className="blue" vertical items={items} />
-        <Menu className="pink" vertical items={items} />
-        <Menu className="green" vertical items={items} />
-      </React.Fragment>
-    )
-  })
-  .add('vertical icons only', () => {
-    const items = [
-      {
-        icon: 'home'
-      },
-      {
-        icon: 'place'
-      },
-      {
-        icon: 'assignment'
-      }
-    ]
-    return (
-      <React.Fragment>
-        <Menu className="blue" vertical icon items={items} />
-        <br />
-        <Menu className="pink" vertical icon items={items} />
-        <br />
-        <Menu className="green" vertical icon items={items} />
-      </React.Fragment>
-    )
-  })
   .add('basic pointing ', () => <Menu pointing items={items} />)
   .add('secondary pointing ', () => <Menu secondary pointing items={items} />)
   .add('basic pointing blue', () => (
